D.S.Reconfigure: generalize type of `reconfigure`
